+++ react/modules/getPaddingRight.ts
@@ -1,0 +1,3 @@
+export default function getPaddingRight(node: HTMLElement) {
+  return parseInt(window.getComputedStyle(node).paddingRight, 10) || 0
+}